Release 1.1.14 (#2466)
# Changelog

## New Features:

- **Teams Revamp**: Announcing a new iteration of Agent teams with the
following features:
- Create a `Team` in one of 3 modes: “Collaborate”, “Coordinate” or
“Route”.
- Various improvements have been made that was wrong with the previous
teams implementation. Including returning structured output from member
agents (for “route” mode), passing images, audio and video to member
agents, etc.
- It has added features like “agentic shared context” between team
members and sharing of individual team member responses with other team
members.
- This also comes with a revamp of Agent and Team debug logs. Use
`debug_mode=True` and `team.print_response(...)` to see it in action.
- Find the docs [here](https://docs.agno.com/teams/introduction). Please
look at the example implementations
[here](https://github.com/agno-agi/agno/blob/c8e47d1643065a0a6ee795c6b063f8576a7a2ef6/cookbook/examples/teams).
- This is the first release. Please give us feedback. Updates and
improvements will follow.
- Support for `Agent(team=[])` is still there, but deprecated (see
below).
- **LiteLLM:** Added [LiteLLM](https://www.litellm.ai/) support, both as
a native implementation and via the `OpenAILike` interface.

## Improvements:

- **Change structured_output to response_format:** Added `use_json_mode:
bool = False` as a parameter of `Agent` and `Team`, which in conjunction
with `response_model=YourModel`, is used to indicate whether the
agent/team model should be forced to respond in json instead of (now
default) structured output. Previous behaviour defaulted to “json-mode”,
but since most models now support native structured output, we are now
defaulting to native structured output. It is now also much simpler to
work with response models, since now only `response_model` needs to be
set. It is not necessary anymore to set `structured_output=True` to
specifically get structured output from the model.

## Bug Fixes:

- **AgentMemory**: Fixed `get_message_pairs()` fetching incorrect
messages.
- **UnionType in Functions**: Fixed issue with function parsing where
pipe-style unions were used in function parameters.

## Deprecations:

- `Agent.structured_output` has been replaced by `Agent.use_json_mode`.
This will be removed in a future major version release.
- `Agent.team` is deprecated with the release of our new Teams
implementation. This will be removed in a future major version release.
