+++ libs/agno/tests/integration/models/openai/chat/test_tool_use.py
@@ -111,7 +111,8 @@ def test_tool_use_with_native_structured_outputs():
         show_tool_calls=True,
         markdown=True,
         response_model=StockPrice,
-        structured_outputs=True,
+        telemetry=False,
+        monitoring=False,
     )
     response = agent.run("What is the current price of TSLA?")
     assert isinstance(response.content, StockPrice)
